+++ example/src/app/app.module.ts
@@ -2,7 +2,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
-import { Ng2ImgFallback } from 'ng2-img-fallback';
+import { Ng2ImgFallbackModule } from 'ng2-img-fallback';
 
 import { AppComponent } from './app.component';
 
@@ -14,7 +14,7 @@ import { AppComponent } from './app.component';
     BrowserModule,
     FormsModule,
     HttpModule,
-    Ng2ImgFallback
+    Ng2ImgFallbackModule
   ],
   providers: [],
   bootstrap: [AppComponent]