+++ examples/ionic/src/app/app.module.ts
@@ -2,7 +2,7 @@ import { NgModule, ErrorHandler } from '@angular/core';
 import { IonicApp, IonicModule, IonicErrorHandler } from 'ionic-angular';
 import { MyApp } from './app.component';
 import { HomePage } from '../pages/home/home';
-import { Ng2ImgFallback } from 'ng2-img-fallback'
+import { Ng2ImgFallbackModule } from 'ng2-img-fallback';
 
 @NgModule({
   declarations: [
@@ -10,8 +10,8 @@ import { Ng2ImgFallback } from 'ng2-img-fallback'
     HomePage
   ],
   imports: [
-    IonicModule.forRoot(MyApp),
-    Ng2ImgFallback
+    Ng2ImgFallbackModule,
+    IonicModule.forRoot(MyApp)
   ],
   bootstrap: [IonicApp],
   entryComponents: [