+++ index.ts
@@ -1,10 +1,5 @@
 /**
  * Created by vadimdez on 04/06/16.
  */
-import { Ng2ImgFallback } from './src/ng2-img-fallback.module';
-
 export * from './src/ng2-img-fallback.module';
-
-export default {
-  directives: [Ng2ImgFallback]
-}
+export * from './src/ng2-img-fallback.directive';